chatty/src/requirements.txt
aiohttp==3.8.3
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.2.0
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
cytoolz==0.12.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.14
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fontawesomefree==6.2.0
frozenlist==1.3.1
hexbytes==0.3.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.6.0.66
parsimonious==0.8.1
protobuf==3.20.3
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
redis==2.10.6
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
semantic-version==2.10.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toolz==0.12.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
varint==1.0.2
web3==5.31.0
websockets==9.1
yarl==1.8.1
zope.interface==5.4.0
